General grammer and spelling cleanup
Also changed the section about being unsure if booting stock with hekate would avoid burning the fuses. Any bootloader that isn't the stock bootloader wont burn fuses, so if you are pushing a payload to boot the system, unless the payload calls directly into the stock bootloader you are safe. 

Left the section about the separate sigpatches for you to do whatever you want with, but the issue has been solved. The issue isn't that hekate needs specifically different sigpatches, its that the sigpatches have to be called differently due to differences in how hekate does patching vs how atmosphere does. This was all required because Tinfoil used to refuse to let you use it if it detected hekate. This has been removed for a while now so the issue isn't something to worry about anymore